partial fix for hussite incident when emperor changes
This is related to EUIV-33333, but contrary to what was said in the
comments of that report, the code is not actually written in a way
which requires the emperor both have -50% warscore and be an OPM. The
reason why it is not working in the save of that bugreport is because
the emperor changed and the previous emperor is not an OPM. But the
previous emperor is the ROOT scope in the incident. This patch fixes the
scope, so that the code works as it is written, but it doesn't fix that
the code is not aligned with the new tooltip. And the warscore condition
is bugged as well, because checking for warscore in is_in_war seems to
not work in some cases. EUIV-35830 is another example of that