make confucian monuments work with harmonized religious groups
they used has_owner_harmonized_religion = yes, but that only works for
eastern religions which are harmonized individually. It doesn't work
if the province has a religion whose group was harmonized. Ideally a
bugfix would change the behavior of has_owner_harmonized_religion, but
that would need a code change, so this fix uses the following workaround
instead:

owner = {
	has_harmonized_with = root
}

Bugreport is at https://forum.paradoxplaza.com/forum/threads/temple-of-confucius-monument-bug.1725911/