turn MLO into a monarchy when FRA forms a PU over them
flavor_fra.109 turns Milan from a vassal into a junior partner without
checking if they are a monarchy which will break the union when they
elect a new ruler. This patch forces them to be a monarchy.

Bugreport: https://forum.paradoxplaza.com/forum/threads/playing-as-france-i-got-a-subjication-cassus-belli-through-an-event-when-the-amrosian-republic-fired-i-forced-them-to-become-my-vassal-through-war.1698229/
and https://forum.paradoxplaza.com/forum/threads/french-event-securing-milanese-lands-personl-union-option-does-not-change-milans-government.1694976/