fix expand infrastructure effect for Expand the Bogue missions
The following missions had the problem that they had the hidden effect
to set a flag which was supposed to increase merchant loyalty when
expanding infrastructure, but that flag was never used. But players
without Leviathan got the alternative effect of 100 dip:
 * mng_expand_bogue
 * mng_expand_bogue2
 * mng_expand_bogue3
 * yua_expand_bogue

This patch also moves the effect out of the hidden effects so that its
custom tooltip is shown. For the other effect of the mission, I changed
the tooltip from mng_pearly_estuary_upgraded_tt to yua_pearly_estuary_upgraded_tt
which was already in use in the mission yua_expand_bogue, because it
explains the actual effects instead of just saying that the estuary will
be upgraded

Bugreport: EUIV-35604 and https://forum.paradoxplaza.com/forum/threads/expand-the-bogue-eoc-mission-is-weird.1718898/