fix or remove outdated dynamic names for the province Dortmund(69)
The province ID 69 used to be called Oberschwaben before 1.30 and
Konstanz before patch 1.12. In 1.30 the province 69 was moved from southern
Germany to northern Germany and renamed to Dortmund. Part of the old
location of province 69 is the new province Constance(4712) since 1.30.
I changed all dynamic names which call 69 a variation
of Constance/Konstanz to 4712 instead(if the culture didn't have a name
for 4712 already) and removed the entries which called 69 a variation of
Oberschwaben.

Bugreport: https://forum.paradoxplaza.com/forum/threads/konstanzs-dynamic-name-in-french-is-in-the-wrong-place.1728238/