Fix encoding in the tutorial
The files were encoded in utf8 instead of windows-1252/cp1252 since
patch 1.37.0. But the files were not properly converted to utf8. Instead
all non-ASCII characters were replaced by the utf8 replacement
character(U+FFFD ; looks like a questionmark in a box) which makes it
impossible to just convert the files back.

So this fix was done by taking the 1.36.2 tutorial saves and applying the
1.37.0 changes which were in lines without the utf8 replacement character. But I might
have missed 1.37.0 changes if they were in lines with that character

Bugreport: EUIV-33125