fix patriarch authority in swedish events/decision
The swedish events "The Support of the Estates"(flavor_swe.108) and
"The Consequences of the Bloodbath of
[Root.Capital.GetName]"(flavor_swe.1080) gave 2000% patriarch authority
if Sweden was orthodox and the decision "Repay Debt to the
Clergy"(swe_repay_estate_church) required 2500% patriarch authority to
remove the modifier from the first event. This patch reduces the gain to
20% which I think was the intended value and the cost to 10% to match
the values for the other religions where the cost is half the gain.

No bugreport, but a reddit post: https://www.reddit.com/r/eu4/comments/1jrkw53/that_might_be_a_bit_difficult/