Fix reconquest war against Kalmyks if the owner doesnt have full cores
When the event "The Arrival of the Kalmyks"(flavor_klm.1) releases the Kalmyk,
the previous owner loses their territorial core which breaks the
reconquest CB from the second event option. This fix adds the
territorial cores back before declaring the war. It also adds a
territorial core on the wargoal 2417 if the owner doesn't have a core on
it.

Bugreport: https://forum.paradoxplaza.com/forum/threads/the-kalmyk-error-and-the-problems-of-building-great-projects.1734253/