Fix burgundian inheritance for non-monarchies
1.37.2 allowed non-monarchies with a royal marriage to be chosen as
bur_strongest_ally, but the event option "[bur_strongest_ally.GetName]
will defend us again."(incidents_bur_inheritance.1.d) is only available
if there also is a monarchy which fulfills the conditions for
bur_strongest_ally. This fix changes the trigger for the event option to
just check if bur_strongest_ally is set and is independent.

Bugreport: https://forum.paradoxplaza.com/forum/threads/burgundian-inheritance-is-still-restricted-to-monarchies-in-1-37-2.1688983/