Do not remove estuary/granary modifiers from flavor_mam.102 if the provinces get conquered
These modifiers replace the default modifiers nile_estuary_modifier or
granary_of_the_mediterranean, which are permanent. The technology in eu4
is not advanced enough to stop the nile from flowing into the sea, so
the estuary modifier can't just be removed. Bringing the original
modifier back would require changes in some on_actions, so I think it is
simpler and clearer to make the new modifiers permanent instead.

Bugreport: https://forum.paradoxplaza.com/forum/threads/expanded-granary-of-the-mediterranean-disappears-when-provinces-change-hands.1701042/