Do not remove "Expanded Daugava Estuary" modifier from Riga if province changes hands
The modifier gets added by the polish mission The Livonian
Lands(plc_break_livonia), and replaces the permanent modifier "Daugava
Estuary"(daugava_estuary_modifier), so it should also be permanent. This
is especially important, because the mission can be completed if the
province is owned by a subject and then the modifier would be lost when
the subject gets integrated.

Bugreport: https://forum.paradoxplaza.com/forum/threads/expanded-daugava-estuary-disappears-completely-after-annexing-vassal.1729695/