Fix temple bonus when switching between reforms
Enacting a reform which has temples_modifier = yes, gives the modifier
gov_expanded_temple_rights_mod to all provinces with a tax building and
removing the reform removes the modifier again. But when switching
between two reforms which have that attribute, it seems that the effect
of the new reform happens before the post_removed_effect/removed_effect
of the removed reform, so the modifier gets removed from all provinces.
This patch fixes it by only removing the modifier if the country does
not have the temples_modifier = yes anymore.

Bugreport: https://forum.paradoxplaza.com/forum/threads/russia-fate-of-the-peasantry-mission-head-of-the-patriarchate-reform-bugs.1674468