Fix devastation trigger and highlighting for "Fortify the Coast" missions
The devastation trigger for the Ming version(mng_fortify_coast) was fixed
to only check coastal provinces, but this change was not applied to the
japanese(mng_fortify_coast2) and korean versions(mng_fortify_coast3).

The highlighting didn't work properly for any case. The Ming version
highlighted provinces which had no devastation and which didn't have a
coastal defense building or 20 dev, even if the country already had 8
provinces to fulfill that part of the condition. This made it hard for
players to find the provinces with devastation. The japanese and korean
versions stopped highlighting when the province had either a coastal
defense or 20 dev and it didn't highlight devastation at all

Bugreport: https://forum.paradoxplaza.com/forum/threads/eu-iv-japanese-mandate-of-heaven-fortify-the-coast-mission-checks-incorrect-provinces-for-devastation.1600486/