Make sure that "A Pretender arises" always has an event option
The only event option in "A Pretender arises"(purple_phoenix.1)
had the trigger that legitimacy must be less than 50, but the event can
also happen if the country has a regency. From the localisation it looks
like there was supposed to be a second option "Support the Pretender"(purple_phoenix.1.b),
which was never implemented. As I don't know what it is supposed to be,
I went with the safe choice and removed the condition from the first
option, so that it is always available.

I think this is the oldest unfixed bug in the game. The bugreport is
from 2014 and the event code is unchanged since the release of the game.

Bugreport: https://forum.paradoxplaza.com/forum/threads/possible-issue-with-event-purple_phoenix-1.775354/