Prevent event "A Zoroastrian Identity" during war
The event "A Zoroastrian Identity"(flavor_per.66) changes the tag to
Eranshahr, which causes bugs with war participation. Blocking the event
from happening during war is a partial fix, because the event could
happen before the war, but the player could click on the event option
after the war started.

Bugreport: https://forum.paradoxplaza.com/forum/threads/forming-eranshar-during-a-war-does-not-keep-trace-of-your-war-participation.1615632/