Fix maritime idea reward from venetian mission "Masters of the Sea"
The reward was wrongly applied when having trade ideas instead of
maritime ideas.

Also prevent giving out the naval ideas reward even when not having
naval ideas.

And simplify the mission effect by moving the tooltips for the effects
out of the if/else blocks which are used for the yes/no tooltip lines

Bugreport: https://forum.paradoxplaza.com/forum/threads/venice-mission-doesnt-give-the-rewards-it-says-it-does.1709353/