make the Kalpi war event chain about Kalpi instead of Bundelkhand
when the event chain was introduced in version 1.8. Kalpi(KLP) did not
exist, so Bundelkhand(BND) was used as a replacement(I guess they owned the
area in question). But Kalpi(KLP) was added in 1.26 and using Bundelkhand(BND)
doesn't make sense anymore.

Bugreport: https://forum.paradoxplaza.com/forum/threads/issues-with-a-jaunpur-event-chain-related-to-kalpi.1616039/