fix removing of prestige penalty when changing secondary religion after completing For the Horde
The mission "For the Horde!"(tatars_devout2) had the effects to remove
the prestige penalty when changing secondary religion and a tooltip
which claimed that it renamed the clergy to shamans. The renaming didn't
work, because they already had that name in the cases when it was
applied, so this was removed in version 1.37.3. But the effect for the
secondary religion penalty used the same flag and was broken as well
even though its tooltip remained.

This patch restores the flag and removes the other uses of the flag
which were related to the renaming of the clergy.

No bugreport, but reddit post: https://www.reddit.com/r/eu4/comments/1j489cm/horde_mission_for_the_horde_is_lying_to_me/