Prevent HSA from switching to generic ideas in flavor_hsa.26
The event "The Crowning of a New Queen" (flavor_hsa.26) which gets
triggered by the mission Queen of the Hansa (hsa_queen_of_the_hansa),
has the option to switch the national ideas to FU2_ideas and the country
name to "The Hanseatic League"(FU2), but this has already happened
if HSA was fomed with the decision "Declare the Hanseatic League"
(declare_hanseatic_league) and taking the ideas again would switch the
country to the next eligible idea set(e.g. generic german ideas with
germanic culture). This patch removes the option to get the ideas if HSA
was formed.

No bugreport, but reddit post: https://www.reddit.com/r/eu4/comments/1o0b4xx/generic_national_ideas_as_hansa/